backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.0.2
openai==1.86.0
opencv-python-headless==4.10.0.84
pillow==11.2.1
proglog==0.1.12
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.23.0
validators
rembg>=2.0.50
librosa>=0.10.0
pydub>=0.25.1
soundfile>=0.12.1
numba>=0.59.0
